+++ src/libs/status.ts
@@ -1,3 +1,12 @@
+import { IconType } from 'react-icons';
+import {
+    RiAlertLine,
+    RiCheckLine,
+    RiIndeterminateCircleLine,
+    RiPlayMiniLine,
+    RiTimeLine,
+} from 'react-icons/ri';
+
 import { AssetType } from '@/modules/common/CommonTypes';
 import { ComputePlanStatus } from '@/modules/computePlans/ComputePlansTypes';
 import { TupleStatus } from '@/modules/tasks/TuplesTypes';
@@ -90,4 +99,75 @@ export const getStatusFromLabel = (
         return getTupleStatusFromLabel(label);
     }
     throw `Unknown asset with status: '${asset}'`;
+};
+
+interface StatusStyle {
+    icon: IconType;
+    tagColor: string;
+    tagBackgroundColor: string;
+    tagSolidColor: string;
+    tagSolidBackgroundColor: string;
+    progressColor: string;
+}
+
+export const getStatusStyle = (
+    status: TupleStatus | ComputePlanStatus
+): StatusStyle => {
+    switch (status) {
+        case TupleStatus.canceled:
+        case ComputePlanStatus.canceled:
+            return {
+                icon: RiIndeterminateCircleLine,
+                tagColor: 'gray.600',
+                tagBackgroundColor: 'gray.100',
+                tagSolidColor: 'white',
+                tagSolidBackgroundColor: 'gray.500',
+                progressColor: 'gray.500',
+            };
+
+        case TupleStatus.waiting:
+        case ComputePlanStatus.waiting:
+        case TupleStatus.todo:
+        case ComputePlanStatus.todo:
+            return {
+                icon: RiTimeLine,
+                tagColor: 'gray.500',
+                tagBackgroundColor: 'gray.50',
+                tagSolidColor: 'white',
+                tagSolidBackgroundColor: 'gray.300',
+                progressColor: 'gray.300',
+            };
+        case TupleStatus.doing:
+        case ComputePlanStatus.doing:
+            return {
+                icon: RiPlayMiniLine,
+                tagColor: 'blue.600',
+                tagBackgroundColor: 'blue.50',
+                tagSolidColor: 'white',
+                tagSolidBackgroundColor: 'blue.500',
+                progressColor: 'blue.500',
+            };
+        case TupleStatus.failed:
+        case ComputePlanStatus.failed:
+            return {
+                icon: RiAlertLine,
+                tagColor: 'red.600',
+                tagBackgroundColor: 'red.50',
+                tagSolidColor: 'white',
+                tagSolidBackgroundColor: 'red.500',
+                progressColor: 'red.500',
+            };
+        case TupleStatus.done:
+        case ComputePlanStatus.done:
+            return {
+                icon: RiCheckLine,
+                tagColor: 'teal.600',
+                tagBackgroundColor: 'teal.50',
+                tagSolidColor: 'white',
+                tagSolidBackgroundColor: 'teal.500',
+                progressColor: 'teal.500',
+            };
+        default:
+            throw 'Unknown status';
+    }
 };